# perf: shallow copy instead of deepcopy