--- a/perception_eval/perception_eval/tool/utils.py
+++ b/perception_eval/perception_eval/tool/utils.py
@@ -14,7 +14,8 @@
 
 from __future__ import annotations
 
-from copy import deepcopy, copy
+from copy import copy
+from copy import deepcopy
 from enum import Enum
 from typing import Any
 from typing import Dict