--- a/perception_eval/perception_eval/tool/utils.py
+++ b/perception_eval/perception_eval/tool/utils.py
@@ -475,8 +475,7 @@ def filter_frame_by_distance(
     Returns:
         PerceptionFrameResult: Filtered frame results.
     """
-    ret_frame = frame.copy_with_shallow_results()
-
+    ret_frame = deepcopy(frame)
     if min_distance is not None:
         min_distance_list = [min_distance] * len(ret_frame.target_labels)
     else: