## Category

- [x] Perception
- [x] Sensing
- [x] Other

## What

Performance optimization, along with the changes in [driving_log_replayer](https://github.com/tier4/driving_log_replayer/pull/581) allows realtime processing (5x speedup) of frames processing.
## Type of change

- [x] Refactoring

## Test performed

Running the driving log replayer and...

- [x] test.sensing_lsim

<details>
<pre>
<code>
[2024-10-30 11:58:31,043] [INFO] [dataset.py:96 load_all_datasets] Start to load dataset ['test/sample_data/']
[2024-10-30 11:58:31,043] [INFO] [dataset.py:105 load_all_datasets] config: load_raw_data: True, evaluation_task: sensing, frame_id: ['base_link']
  0%|                                                                                                                                       | 0/1 [00:00<?, ?it/s][2024-10-30 11:58:31,062] [WARNING] [label.py:256 convert_label] Label human.pedestrian.adult is not registered.
100%|███████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 1/1 [00:00<00:00, 56.11it/s]
[2024-10-30 11:58:31,062] [INFO] [dataset.py:120 load_all_datasets] Finish loading dataset
bicycle (-> bicycle): 0 
vehicle.bicycle (-> bicycle): 0 
car (-> car): 0 
vehicle.car (-> car): 0 
vehicle.construction (-> car): 0 
vehicle.emergency (ambulance & police) (-> car): 0 
vehicle.police (-> car): 0 
vehicle.fire (-> car): 0 
vehicle.ambulance (-> car): 0 
pedestrian (-> pedestrian): 0 
stroller (-> pedestrian): 0 
pedestrian.adult (-> pedestrian): 0 
pedestrian.child (-> pedestrian): 0 
pedestrian.construction_worker (-> pedestrian): 0 
pedestrian.personal_mobility (-> pedestrian): 0 
pedestrian.police_officer (-> pedestrian): 0 
pedestrian.stroller (-> pedestrian): 0 
pedestrian.wheelchair (-> pedestrian): 0 
construction_worker (-> pedestrian): 0 
animal (-> unknown): 0 
unknown (-> unknown): 0 
movable_object.barrier (-> unknown): 0 
movable_object.debris (-> unknown): 0 
movable_object.pushable_pullable (-> unknown): 0 
movable_object.trafficcone (-> unknown): 0 
movable_object.traffic_cone (-> unknown): 0 
static_object.bicycle rack (-> unknown): 0 
static_object.bollard (-> unknown): 0 
forklift (-> unknown): 0 
false_positive (-> false_positive): 0 
bus (-> bus): 0 
vehicle.bus (bendy & rigid) (-> bus): 0 
vehicle.bus (-> bus): 0 
truck (-> truck): 0 
vehicle.truck (-> truck): 0 
trailer (-> truck): 0 
vehicle.trailer (-> truck): 0 
motorbike (-> motorbike): 0 
motorcycle (-> motorbike): 0 
vehicle.motorcycle (-> motorbike): 0 

[2024-10-30 11:58:31,124] [INFO] [sensing_lsim.py:141 visualize] all detections were succeeded.
[2024-10-30 11:58:31,124] [INFO] [sensing_lsim.py:144 visualize] [SUCCESS] Inside points: 1, Is detected: True, Nearest point: [1.8647348e+01 5.9553642e+01 5.0377835e-02]

[2024-10-30 11:58:31,124] [WARNING] [sensing_lsim.py:118 get_final_result] 0 fail results.
[2024-10-30 11:58:31,124] [INFO] [sensing_lsim.py:209 <module>] Frame result example (frame_results[0]: 1 success, 0 fail
[2024-10-30 11:58:31,124] [INFO] [sensing_lsim.py:215 <module>] Failed to be Non-detected pointcloud example (frame_results[0]): 0
[2024-10-30 11:58:31,124] [INFO] [sensing_lsim.py:221 <module>] Start visualizing sensing results
Visualize results for each frame: 100%|█████████████████████████████████████████████████████████████████████████████████████████████| 1/1 [00:00<00:00, 22.47it/s
</code>
</pre>
</details>

- [x] test.perception_lsim

<details>
<pre>
<code>
==================================================Start Detection==================================================
[2024-10-30 11:57:07,600] [INFO] [dataset.py:96 load_all_datasets] Start to load dataset ['test/sample_data/']
[2024-10-30 11:57:07,600] [INFO] [dataset.py:105 load_all_datasets] config: load_raw_data: True, evaluation_task: detection, frame_id: ['base_link']
  0%|                                                                                                                                       | 0/1 [00:00<?, ?it/s][2024-10-30 11:57:07,621] [WARNING] [label.py:256 convert_label] Label human.pedestrian.adult is not registered.
100%|███████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 1/1 [00:00<00:00, 51.59it/s]
[2024-10-30 11:57:07,621] [INFO] [dataset.py:120 load_all_datasets] Finish loading dataset
bicycle (-> bicycle): 1 
vehicle.bicycle (-> bicycle): 0 
car (-> car): 1 
vehicle.car (-> car): 0 
vehicle.construction (-> car): 0 
vehicle.emergency (ambulance & police) (-> car): 0 
vehicle.police (-> car): 0 
vehicle.fire (-> car): 0 
vehicle.ambulance (-> car): 0 
pedestrian (-> pedestrian): 1 
stroller (-> pedestrian): 0 
pedestrian.adult (-> pedestrian): 0 
pedestrian.child (-> pedestrian): 0 
pedestrian.construction_worker (-> pedestrian): 0 
pedestrian.personal_mobility (-> pedestrian): 0 
pedestrian.police_officer (-> pedestrian): 0 
pedestrian.stroller (-> pedestrian): 0 
pedestrian.wheelchair (-> pedestrian): 0 
construction_worker (-> pedestrian): 0 
animal (-> unknown): 0 
unknown (-> unknown): 0 
movable_object.barrier (-> unknown): 0 
movable_object.debris (-> unknown): 0 
movable_object.pushable_pullable (-> unknown): 0 
movable_object.trafficcone (-> unknown): 0 
movable_object.traffic_cone (-> unknown): 0 
static_object.bicycle rack (-> unknown): 0 
static_object.bollard (-> unknown): 0 
forklift (-> unknown): 0 
false_positive (-> false_positive): 0 
bus (-> bus): 0 
vehicle.bus (bendy & rigid) (-> bus): 0 
vehicle.bus (-> bus): 0 
truck (-> truck): 0 
vehicle.truck (-> truck): 0 
trailer (-> truck): 0 
vehicle.trailer (-> truck): 0 
motorbike (-> motorbike): 1 
motorcycle (-> motorbike): 0 
vehicle.motorcycle (-> motorbike): 0 

[2024-10-30 11:57:07,717] [INFO] [perception_lsim.py:153 visualize] 0 TP objects, 0 FP objects, 0 FN objects
[2024-10-30 11:57:07,717] [INFO] [perception_lsim.py:142 get_final_result] Number of fails for critical objects: 0
[2024-10-30 11:57:07,718] [INFO] [perception_lsim.py:146 get_final_result] final metrics result 

Frame:
 Total Num: 1
 Skipped Frames: []
 Skipped Frames Count: 0

Ground Truth Num: 0

mAP: inf, mAPH: inf (Center Distance)

|      Label | car(1.0) |  bicycle(1.0) |  pedestrian(1.0) |  motorbike(1.0) | 
| :--------: | :---: | :---: | :---: | :---: |
| Predict_num | 0 | 0 | 0 | 0 |
|         AP | inf |  inf |  inf |  inf | 
|        APH | inf |  inf |  inf |  inf | 

mAP: inf, mAPH: inf (Center Distance)

|      Label | car(2.0) |  bicycle(2.0) |  pedestrian(2.0) |  motorbike(2.0) | 
| :--------: | :---: | :---: | :---: | :---: |
| Predict_num | 0 | 0 | 0 | 0 |
|         AP | inf |  inf |  inf |  inf | 
|        APH | inf |  inf |  inf |  inf | 

mAP: inf, mAPH: inf (IoU 2D)

|      Label | car(0.5) |  bicycle(0.5) |  pedestrian(0.5) |  motorbike(0.5) | 
| :--------: | :---: | :---: | :---: | :---: |
| Predict_num | 0 | 0 | 0 | 0 |
|         AP | inf |  inf |  inf |  inf | 
|        APH | inf |  inf |  inf |  inf | 

mAP: inf, mAPH: inf (IoU 3D)

|      Label | car(0.5) |  bicycle(0.5) |  pedestrian(0.5) |  motorbike(0.5) | 
| :--------: | :---: | :---: | :---: | :---: |
| Predict_num | 0 | 0 | 0 | 0 |
|         AP | inf |  inf |  inf |  inf | 
|        APH | inf |  inf |  inf |  inf | 

mAP: inf, mAPH: inf (Plane Distance)

|      Label | car(2.0) |  bicycle(2.0) |  pedestrian(2.0) |  motorbike(2.0) | 
| :--------: | :---: | :---: | :---: | :---: |
| Predict_num | 0 | 0 | 0 | 0 |
|         AP | inf |  inf |  inf |  inf | 
|        APH | inf |  inf |  inf |  inf | 

mAP: inf, mAPH: inf (Plane Distance)

|      Label | car(3.0) |  bicycle(3.0) |  pedestrian(3.0) |  motorbike(3.0) | 
| :--------: | :---: | :---: | :---: | :---: |
| Predict_num | 0 | 0 | 0 | 0 |
|         AP | inf |  inf |  inf |  inf | 
|        APH | inf |  inf |  inf |  inf | 

[2024-10-30 11:57:07,718] [INFO] [perception_lsim.py:210 <module>] Frame result example (frame_results[0]): 
{'frame_ground_truth': {'frame_name': '0',
                        'objects': [],
                        'raw_data': {<FrameID.CAM_FRONT: 'cam_front'>: ' --- length of element 900 ---,',
                                     <FrameID.LIDAR_TOP: 'lidar_top'>: ' --- length of element 34688 ---,'},
                        'transforms': ' --- length of element 25 ---,',
                        'unix_time': 1532402927647951},
 'frame_name': '0',
 'metrics_score': {'classification_config': None,
                   'classification_scores': [],
                   'detection_config': {'center_distance_thresholds': ' --- length of element 2 ---,',
                                        'iou_2d_thresholds': [' --- length of element 4 ---,'],
                                        'iou_3d_thresholds': [' --- length of element 4 ---,'],
                                        'plane_distance_thresholds': ' --- length of element 2 ---,',
                                        'target_labels': ' --- length of element 4 ---,'},
                   'evaluation_task': 'detection',
                   'maps': ' --- length of element 6 ---,',
                   'prediction_config': None,
                   'prediction_scores': [],
                   'tracking_config': None,
                   'tracking_scores': []},
 'object_results': [],
 'pass_fail_result': {'critical_object_filter_config': {'confidence_threshold_list': None,
                                                        'filtering_params': {'confidence_threshold_list': None,
                                                                             'ignore_attributes': ['cycle_state.without_rider'],
                                                                             'max_distance_list': None,
                                                                             'max_x_position_list': ' --- length of '
                                                                                                    'element 4 ---,',
                                                                             'max_y_position_list': ' --- length of '
                                                                                                    'element 4 ---,',
                                                                             'min_distance_list': None,
                                                                             'min_point_numbers': None,
                                                                             'target_labels': ' --- length of element '
                                                                                              '4 ---,',
                                                                             'target_uuids': None},
                                                        'ignore_attributes': ['cycle_state.without_rider'],
                                                        'max_distance_list': None,
                                                        'max_x_position_list': ' --- length of element 4 ---,',
                                                        'max_y_position_list': ' --- length of element 4 ---,',
                                                        'min_distance_list': None,
                                                        'min_point_numbers': None,
                                                        'target_labels': ' --- length of element 4 ---,',
                                                        'target_uuids': None},
                      'fn_objects': [],
                      'fp_object_results': [],
                      'frame_number': '0',
                      'frame_pass_fail_config': {'confidence_threshold_list': None,
                                                 'evaluation_task': 'detection',
                                                 'matching_threshold_list': ' --- length of element 4 ---,',
                                                 'target_labels': ' --- length of element 4 ---,'},
                      'tn_objects': [],
                      'tp_object_results': [],
                      'transforms': ' --- length of element 25 ---,',
                      'unix_time': 1532402927647951},
 'target_labels': ' --- length of element 4 ---,',
 'unix_time': 1532402927647951}

[2024-10-30 11:57:07,719] [INFO] [perception_lsim.py:222 <module>] Detection Metrics example (final_metric_score): 
{'classification_config': None,
 'classification_scores': [],
 'detection_config': {'center_distance_thresholds': [[1.0, 1.0, 1.0, 1.0], [2.0, 2.0, 2.0, 2.0]],
                      'iou_2d_thresholds': [[0.5, 0.5, 0.5, 0.5]],
                      'iou_3d_thresholds': [[0.5, 0.5, 0.5, 0.5]],
                      'plane_distance_thresholds': [[2.0, 2.0, 2.0, 2.0], [3.0, 3.0, 3.0, 3.0]],
                      'target_labels': ['car', 'bicycle', 'pedestrian', 'motorbike']},
 'evaluation_task': 'detection',
 'maps': ' --- length of element 6 ---,',
 'prediction_config': None,
 'prediction_scores': [],
 'tracking_config': None,
 'tracking_scores': []}

[2024-10-30 11:57:07,719] [INFO] [perception_lsim.py:228 <module>] mAP result example (final_metric_score.maps[0].aps[0]): 
{'aphs': [{'ap': inf,
           'fp_list': [],
           'matching_average': None,
           'matching_mode': 'Center Distance',
           'matching_standard_deviation': None,
           'matching_threshold_list': [1.0],
           'num_ground_truth': 0,
           'objects_results_num': 0,
           'target_labels': ['car'],
           'tp_list': [],
           'tp_metrics': {}},
          {'ap': inf,
           'fp_list': [],
           'matching_average': None,
           'matching_mode': 'Center Distance',
           'matching_standard_deviation': None,
           'matching_threshold_list': [1.0],
           'num_ground_truth': 0,
           'objects_results_num': 0,
           'target_labels': ['bicycle'],
           'tp_list': [],
           'tp_metrics': {}},
          {'ap': inf,
           'fp_list': [],
           'matching_average': None,
           'matching_mode': 'Center Distance',
           'matching_standard_deviation': None,
           'matching_threshold_list': [1.0],
           'num_ground_truth': 0,
           'objects_results_num': 0,
           'target_labels': ['pedestrian'],
           'tp_list': [],
           'tp_metrics': {}},
          {'ap': inf,
           'fp_list': [],
           'matching_average': None,
           'matching_mode': 'Center Distance',
           'matching_standard_deviation': None,
           'matching_threshold_list': [1.0],
           'num_ground_truth': 0,
           'objects_results_num': 0,
           'target_labels': ['motorbike'],
           'tp_list': [],
           'tp_metrics': {}}],
 'aps': [{'ap': inf,
          'fp_list': [],
          'matching_average': None,
          'matching_mode': 'Center Distance',
          'matching_standard_deviation': None,
          'matching_threshold_list': [1.0],
          'num_ground_truth': 0,
          'objects_results_num': 0,
          'target_labels': ['car'],
          'tp_list': [],
          'tp_metrics': {}},
         {'ap': inf,
          'fp_list': [],
          'matching_average': None,
          'matching_mode': 'Center Distance',
          'matching_standard_deviation': None,
          'matching_threshold_list': [1.0],
          'num_ground_truth': 0,
          'objects_results_num': 0,
          'target_labels': ['bicycle'],
          'tp_list': [],
          'tp_metrics': {}},
         {'ap': inf,
          'fp_list': [],
          'matching_average': None,
          'matching_mode': 'Center Distance',
          'matching_standard_deviation': None,
          'matching_threshold_list': [1.0],
          'num_ground_truth': 0,
          'objects_results_num': 0,
          'target_labels': ['pedestrian'],
          'tp_list': [],
          'tp_metrics': {}},
         {'ap': inf,
          'fp_list': [],
          'matching_average': None,
          'matching_mode': 'Center Distance',
          'matching_standard_deviation': None,
          'matching_threshold_list': [1.0],
          'num_ground_truth': 0,
          'objects_results_num': 0,
          'target_labels': ['motorbike'],
          'tp_list': [],
          'tp_metrics': {}}],
 'is_detection_2d': False,
 'map': inf,
 'maph': inf,
 'matching_mode': 'Center Distance',
 'matching_threshold_list': [1.0, 1.0, 1.0, 1.0],
 'target_labels': ['car', 'bicycle', 'pedestrian', 'motorbike']}

[2024-10-30 11:57:07,719] [INFO] [perception_lsim.py:235 <module>] Start visualizing detection results
Visualize results for each frame: 100%|█████████████████████████████████████████████████████████████████████████████████████████████| 1/1 [00:00<00:00, 22.12it/s]
Updating DataFrame: 100%|██████████████████████████████████████████████████████████████████████████████████████████████████████████| 1/1 [00:00<00:00, 737.65it/s]
[2024-10-30 11:57:07,785] [WARNING] [perception_analyzer3d.py:415 analyze] There is no DataFrame to be able to analyze.
==================================================Start Tracking==================================================
[2024-10-30 11:57:07,785] [INFO] [dataset.py:96 load_all_datasets] Start to load dataset ['test/sample_data/']
[2024-10-30 11:57:07,785] [INFO] [dataset.py:105 load_all_datasets] config: load_raw_data: True, evaluation_task: tracking, frame_id: ['map']
  0%|                                                                                                                                       | 0/1 [00:00<?, ?it/s][2024-10-30 11:57:07,802] [WARNING] [label.py:256 convert_label] Label human.pedestrian.adult is not registered.
100%|███████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 1/1 [00:00<00:00, 63.44it/s]
[2024-10-30 11:57:07,803] [INFO] [dataset.py:120 load_all_datasets] Finish loading dataset
bicycle (-> bicycle): 1 
vehicle.bicycle (-> bicycle): 0 
car (-> car): 1 
vehicle.car (-> car): 0 
vehicle.construction (-> car): 0 
vehicle.emergency (ambulance & police) (-> car): 0 
vehicle.police (-> car): 0 
vehicle.fire (-> car): 0 
vehicle.ambulance (-> car): 0 
pedestrian (-> pedestrian): 1 
stroller (-> pedestrian): 0 
pedestrian.adult (-> pedestrian): 0 
pedestrian.child (-> pedestrian): 0 
pedestrian.construction_worker (-> pedestrian): 0 
pedestrian.personal_mobility (-> pedestrian): 0 
pedestrian.police_officer (-> pedestrian): 0 
pedestrian.stroller (-> pedestrian): 0 
pedestrian.wheelchair (-> pedestrian): 0 
construction_worker (-> pedestrian): 0 
animal (-> unknown): 0 
unknown (-> unknown): 0 
movable_object.barrier (-> unknown): 0 
movable_object.debris (-> unknown): 0 
movable_object.pushable_pullable (-> unknown): 0 
movable_object.trafficcone (-> unknown): 0 
movable_object.traffic_cone (-> unknown): 0 
static_object.bicycle rack (-> unknown): 0 
static_object.bollard (-> unknown): 0 
forklift (-> unknown): 0 
false_positive (-> false_positive): 0 
bus (-> bus): 0 
vehicle.bus (bendy & rigid) (-> bus): 0 
vehicle.bus (-> bus): 0 
truck (-> truck): 0 
vehicle.truck (-> truck): 0 
trailer (-> truck): 0 
vehicle.trailer (-> truck): 0 
motorbike (-> motorbike): 1 
motorcycle (-> motorbike): 0 
vehicle.motorcycle (-> motorbike): 0 

[2024-10-30 11:57:07,840] [INFO] [dataset.py:340 get_interpolated_now_frame] Only before frame is available for interpolation
…