@ralwing 2024-11-08
I've just noticed, that there is a bug: for more than one criterion, it only returns the last one.