@ralwing 2024-11-20
@YoshiRi 
Ok, I've double tested this solution, and seems to be ready to review.