@ralwing 2024-11-22
There is much bigger problem with this code, than i previously though.
The `estimated_object` and `ground_truth_object` are copied multiple times in the code, in fact the class `DynamicObjectWithPerceptionResult` has a few copies of it under the hood. So when there is a call to [deepcopy(perceptionframeresult) ](https://github.com/tier4/autoware_perception_evaluation/blob/968d12f29254da8a1691297401c5a19262fcca9e/perception_eval/perception_eval/tool/utils.py#L321) these estimated_object ground_truth_object are unnecessarily copied multiple times. e.g. every [MatchingMethod](https://github.com/tier4/autoware_perception_evaluation/blob/968d12f29254da8a1691297401c5a19262fcca9e/perception_eval/perception_eval/evaluation/matching/object_matching.py#L113) implementation contains a copy of them.