@ralwing 2024-11-25
I hope this time I have considered all cases, however I have a request that T4 checks this PR on your data. This solution is not 2x speedup, but still it significantly shortens the processing time.  
Thanks in advance.